Finished moving pathfinding related classes in PathFinding
